fixed thumb generation: -added checking for thumb creation; -getting frame by microseconds not milliseconds.

+++ app/platforms/android/src/com/noplanbees/tbm/IntentHandler.java
@@ -155,7 +155,9 @@ public class IntentHandler {
 		friend.updateStatus(intent);
 
 		if (status == Video.IncomingVideoStatus.DOWNLOADED) {
-			friend.createThumb(videoId);
+			if(!friend.createThumb(videoId)){
+               return;
+            }
 
 			friend.deleteAllViewedVideos();
